+++ src/utils/types.ts
@@ -1,2 +1,2 @@
-export type TicketStatus = 'created' | 'opened' | 'closed' | 'error';
+export type TicketStatus = 'created' | 'opened' | 'closed' | 'adminOnly' | 'error';
 export type SavedRoleTypes = 'staff' | 'admin';