+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# Dev Update 1.0.2
+# Dev Update v1.0.3
+- Utility function for extracting role ids from the database
+- Added 'Admin Only' button next to ticket close button
+- Changing a bunch of files to use the lang json
+- Added 'Ticket Category' to ticket config database
+- Added ticket category adding command
+- Made sure logic was fixed for previous ticket category method to current
+- Fixed styling with the Admin Only and Close Ticket buttons
+
+# Dev Update v1.0.2
 - Changelog finally lol
 - Adding admin/staff roles v2 (add-role slash command)
 - Removing admin/staff roles (remove-role slash command)